+++ Assets/VRM10_Samples/VRM10Viewer/UIToolkit.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed0cdb4edf7f31e40bb922d4103ade0e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
