+++ Assets/VRM10_Samples/VRM10Viewer/uGui/EmotionFields.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9f83ba27c7836e4290085e6557ac2bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
