+++ Assets/VRM10_Samples/VRM10Viewer/uGui/ObjectMap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e59a4c52333b64c40b4954ba292ddb28
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
