+++ Assets/VRM10_Samples/VRM10Viewer/uGui/TextFields.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f642f3373e8ad24082bbe4b4aac6c62
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
