+++ Assets/VRM10_Samples/VRM10Viewer/uGui/UIFields.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1508380c7e603d54d9d356868a357b20
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
