+++ Assets/VRM10_Samples/VRM10Viewer/uGui/VRM10ViewerUI.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using UniGLTF;
+﻿using UniGLTF;
 using UniGLTF.SpringBoneJobs.Blittables;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.UI;
-
 
 namespace UniVRM10.VRM10Viewer
 {
@@ -73,73 +69,7 @@ namespace UniVRM10.VRM10Viewer
 
         [SerializeField, Header("expression")]
         Toggle m_enableAutoExpression = default;
-        [Serializable]
-        class EmotionFields
-        {
-            public Slider m_expression;
-            public Toggle m_binary;
-            public bool m_useOverride;
-            public Dropdown m_overrideMouth;
-            public Dropdown m_overrideBlink;
-            public Dropdown m_overrideLookAt;
 
-            public void Reset(ObjectMap map, string name, bool useOveride)
-            {
-                m_expression = map.Get<Slider>($"Slider{name}");
-                m_binary = map.Get<Toggle>($"Binary{name}");
-                m_useOverride = useOveride;
-                if (useOveride)
-                {
-                    m_overrideMouth = map.Get<Dropdown>($"Override{name}Mouth");
-                    m_overrideBlink = map.Get<Dropdown>($"Override{name}Blink");
-                    m_overrideLookAt = map.Get<Dropdown>($"Override{name}LookAt");
-                }
-            }
-
-            static int GetOverrideIndex(UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType value)
-            {
-                switch (value)
-                {
-                    case UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType.none: return 0;
-                    case UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType.block: return 1;
-                    case UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType.blend: return 2;
-                    default: return -1;
-                }
-            }
-
-            static UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType ToOverrideType(int index)
-            {
-                switch (index)
-                {
-                    case 0: return UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType.none;
-                    case 1: return UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType.block;
-                    case 2: return UniGLTF.Extensions.VRMC_vrm.ExpressionOverrideType.blend;
-                    default: throw new ArgumentException();
-                }
-            }
-
-            public void OnLoad(VRM10Expression expression)
-            {
-                m_binary.isOn = expression.IsBinary;
-                if (m_useOverride)
-                {
-                    m_overrideMouth.SetValueWithoutNotify(GetOverrideIndex(expression.OverrideMouth));
-                    m_overrideBlink.SetValueWithoutNotify(GetOverrideIndex(expression.OverrideBlink));
-                    m_overrideLookAt.SetValueWithoutNotify(GetOverrideIndex(expression.OverrideLookAt));
-                }
-            }
-
-            public void ApplyRuntime(VRM10Expression expression)
-            {
-                expression.IsBinary = m_binary.isOn;
-                if (m_useOverride)
-                {
-                    expression.OverrideMouth = ToOverrideType(m_overrideMouth.value);
-                    expression.OverrideBlink = ToOverrideType(m_overrideBlink.value);
-                    expression.OverrideLookAt = ToOverrideType(m_overrideLookAt.value);
-                }
-            }
-        }
         [SerializeField]
         EmotionFields m_happy;
         [SerializeField]
@@ -178,180 +108,16 @@ namespace UniVRM10.VRM10Viewer
         [SerializeField]
         Slider m_pitch = default;
 
-        [Serializable]
-        class TextFields
-        {
-            [SerializeField]
-            Text m_textModelTitle = default;
-            [SerializeField]
-            Text m_textModelVersion = default;
-            [SerializeField]
-            Text m_textModelAuthor = default;
-            [SerializeField]
-            Text m_textModelCopyright = default;
-            [SerializeField]
-            Text m_textModelContact = default;
-            [SerializeField]
-            Text m_textModelReference = default;
-            [SerializeField]
-            RawImage m_thumbnail = default;
-
-            [SerializeField, Header("CharacterPermission")]
-            Text m_textPermissionAllowed = default;
-            [SerializeField]
-            Text m_textPermissionViolent = default;
-            [SerializeField]
-            Text m_textPermissionSexual = default;
-            [SerializeField]
-            Text m_textPermissionCommercial = default;
-            [SerializeField]
-            Text m_textPermissionOther = default;
-
-            [SerializeField, Header("DistributionLicense")]
-            Text m_textDistributionLicense = default;
-            [SerializeField]
-            Text m_textDistributionOther = default;
-
-            public void Reset(ObjectMap map)
-            {
-                m_textModelTitle = map.Get<Text>("Title (1)");
-                m_textModelVersion = map.Get<Text>("Version (1)");
-                m_textModelAuthor = map.Get<Text>("Author (1)");
-                m_textModelCopyright = map.Get<Text>("Copyright (1)");
-                m_textModelContact = map.Get<Text>("Contact (1)");
-                m_textModelReference = map.Get<Text>("Reference (1)");
-
-                m_textPermissionAllowed = map.Get<Text>("AllowedUser (1)");
-                m_textPermissionViolent = map.Get<Text>("Violent (1)");
-                m_textPermissionSexual = map.Get<Text>("Sexual (1)");
-                m_textPermissionCommercial = map.Get<Text>("Commercial (1)");
-                m_textPermissionOther = map.Get<Text>("Other (1)");
-
-                m_textDistributionLicense = map.Get<Text>("LicenseType (1)");
-                m_textDistributionOther = map.Get<Text>("OtherLicense (1)");
-
-#if UNITY_2022_3_OR_NEWER
-                var images = GameObject.FindObjectsByType<RawImage>(FindObjectsSortMode.InstanceID);
-#else
-                var images = GameObject.FindObjectsOfType<RawImage>();
-#endif
-                m_thumbnail = images.First(x => x.name == "RawImage");
-            }
-
-            public void Start()
-            {
-                m_textModelTitle.text = "";
-                m_textModelVersion.text = "";
-                m_textModelAuthor.text = "";
-                m_textModelCopyright.text = "";
-                m_textModelContact.text = "";
-                m_textModelReference.text = "";
-
-                m_textPermissionAllowed.text = "";
-                m_textPermissionViolent.text = "";
-                m_textPermissionSexual.text = "";
-                m_textPermissionCommercial.text = "";
-                m_textPermissionOther.text = "";
-
-                m_textDistributionLicense.text = "";
-                m_textDistributionOther.text = "";
-            }
-
-            public void UpdateMeta(Texture2D thumbnail, UniGLTF.Extensions.VRMC_vrm.Meta meta, Migration.Vrm0Meta meta0)
-            {
-                m_thumbnail.texture = thumbnail;
-
-                if (meta != null)
-                {
-                    m_textModelTitle.text = meta.Name;
-                    m_textModelVersion.text = meta.Version;
-                    m_textModelAuthor.text = meta.Authors[0];
-                    m_textModelCopyright.text = meta.CopyrightInformation;
-                    m_textModelContact.text = meta.ContactInformation;
-                    if (meta.References != null && meta.References.Count > 0)
-                    {
-                        m_textModelReference.text = meta.References[0];
-                    }
-                    m_textPermissionAllowed.text = meta.AvatarPermission.ToString();
-                    m_textPermissionViolent.text = meta.AllowExcessivelyViolentUsage.ToString();
-                    m_textPermissionSexual.text = meta.AllowExcessivelySexualUsage.ToString();
-                    m_textPermissionCommercial.text = meta.CommercialUsage.ToString();
-                    // m_textPermissionOther.text = meta.OtherPermissionUrl;
-
-                    // m_textDistributionLicense.text = meta.ModificationLicense.ToString();
-                    m_textDistributionOther.text = meta.OtherLicenseUrl;
-                }
-
-                if (meta0 != null)
-                {
-                    m_textModelTitle.text = meta0.title;
-                    m_textModelVersion.text = meta0.version;
-                    m_textModelAuthor.text = meta0.author;
-                    m_textModelContact.text = meta0.contactInformation;
-                    m_textModelReference.text = meta0.reference;
-                    m_textPermissionAllowed.text = meta0.allowedUser.ToString();
-                    m_textPermissionViolent.text = meta0.violentUsage.ToString();
-                    m_textPermissionSexual.text = meta0.sexualUsage.ToString();
-                    m_textPermissionCommercial.text = meta0.commercialUsage.ToString();
-                    m_textPermissionOther.text = meta0.otherPermissionUrl;
-                    // m_textDistributionLicense.text = meta0.ModificationLicense.ToString();
-                    m_textDistributionOther.text = meta0.otherLicenseUrl;
-                }
-            }
-        }
         [SerializeField]
         TextFields m_texts = default;
 
-        [Serializable]
-        class UIFields
-        {
-            [SerializeField]
-            Toggle ToggleMotionTPose = default;
-
-            [SerializeField]
-            Toggle ToggleMotionBVH = default;
-
-            [SerializeField]
-            ToggleGroup ToggleMotion = default;
-
-            public void Reset(ObjectMap map)
-            {
-                ToggleMotionTPose = map.Get<Toggle>("TPose");
-                ToggleMotionBVH = map.Get<Toggle>("BVH");
-                ToggleMotion = map.Get<ToggleGroup>("_Motion_");
-            }
-
-            public bool IsTPose
-            {
-                get => ToggleMotion.ActiveToggles().FirstOrDefault() == ToggleMotionTPose;
-                set
-                {
-                    ToggleMotionTPose.isOn = value;
-                    ToggleMotionBVH.isOn = !value;
-                }
-            }
-        }
         [SerializeField]
         UIFields m_ui = default;
 
-        class ObjectMap
-        {
-            Dictionary<string, GameObject> _map = new();
-            public IReadOnlyDictionary<string, GameObject> Objects => _map;
-
-            public ObjectMap(GameObject root)
-            {
-                foreach (var x in root.GetComponentsInChildren<Transform>())
-                {
-                    _map[x.name] = x.gameObject;
-                }
-            }
-
-            public T Get<T>(string name) where T : Component
-            {
-                return _map[name].GetComponent<T>();
-            }
-        }
+        VRM10ViewerController m_controller;
+        VRM10AutoExpression m_autoEmotion;
+        VRM10Blinker m_autoBlink;
+        VRM10AIUEO m_autoLipsync;
 
         private void Reset()
         {
@@ -403,11 +169,6 @@ namespace UniVRM10.VRM10Viewer
             m_lookAtTarget = GameObject.FindObjectOfType<VRM10TargetMover>().gameObject;
 #endif
         }
-
-        VRM10ViewerController m_controller;
-        VRM10AutoExpression m_autoEmotion;
-        VRM10Blinker m_autoBlink;
-        VRM10AIUEO m_autoLipsync;
 
         void OnLoaded(Loaded loaded)
         {
